[PATRO23-7] create KeycloakService
